feat: add EditorialService interface
New layer of abstraction, encapsulation, flexibility and separation of
concerns.
